Refactor integration tests into modular fixtures and scenarios
- Extract factory methods to __tests__/test-fixtures.ts with overrides pattern
- Create separate scenario files organized by complexity:
  - basic-scenarios.ts: movement, variables, conditions, transforms
  - multi-actor-scenarios.ts: collision and interaction tests
  - chase-game-scenario.ts: keyboard input + AI movement
  - score-game-scenario.ts: global/actor variable accumulation
- Simplify main test file to import and run scenarios
- Add TestScenario type and runScenario() helper for consistent test structure

This makes tests more maintainable and easier to add new scenarios.
